+++ framework/src/transformers/index.js
@@ -3,7 +3,7 @@ import get from 'lodash-es/get.js'
 import { defu as merge } from 'defu'
 
 import core from './core.js'
-import comb from './comb.js'
+import purge from './purge.js'
 import sixHex from './sixHex.js'
 import minify from './minify.js'
 import baseUrl from './baseUrl.js'
@@ -162,9 +162,10 @@ export async function run(html = '', config = {}) {
    *
    * Add a base URL to relative paths.
    */
-  if (get(config, 'baseURL', get(config, 'baseUrl'))) {
+  const baseConfig = get(config, 'baseURL', get(config, 'baseUrl'))
+  if (baseConfig) {
     posthtmlPlugins.push(
-      baseUrl(get(config, 'baseURL', get(config, 'baseUrl', {})))
+      baseUrl(baseConfig)
     )
   }
 
@@ -207,7 +208,7 @@ export async function run(html = '', config = {}) {
    * Remove unused CSS, uglify classes etc.
    */
   if (get(config, 'css.purge')) {
-    posthtmlPlugins.push(comb(config.css.purge))
+    posthtmlPlugins.push(purge(config.css.purge))
   }
 
   /**